rework the usage of page headers
These changes make it possible to use the library on systems
which don't support multiple writes to the same flash address
as long as the new value is smaller than the previous one (1 -> 0 transition)
such as STM32L4. Page header is now written in 32 bytes, split across
4 64-bit values.

Factory page copying is also reworked - now the library always scans the
entire factory page in init() and copies entries which aren't found in
the current page from the factory page. This avoids the need to reformat
the entire page just to copy new data from the factory page.